Add files via upload
The Power BI Dashboard, titled “Analysis of Mobile 2023”, is a comprehensive infographic that presents a variety of statistics and insights about mobile phones. Here are the principal details:

•	Total Mobile: Data from 600 mobile phones were analyzed.
•	Brands: The brands include Apple, Infinix, JioPhone, Mi, Motorola, One Plus, Realme, Redmi, Samsung, Vivo, and Xiaomi.
•	Operating System of Mobile: A bar graph shows that Android is the most used operating system with 453 phones.
•	Mobile Battery Capacity: A pie chart displays the distribution of battery capacities among the analyzed phones.
•	Analysis of 5G availability: A horizontal bar graph indicates that most phones do not specify this feature.
•	Analysis of mobile brands: A pie chart breaks down the percentage representation of each brand in the analysis.
•	Internal Storage: A vertical bar graph compares different internal storage capacities available in these mobiles.
•	RAM Capacity: A pie chart represents various RAM capacities found in these devices.

This infographic seems to be a comprehensive analysis of mobile phones in the year 2023. It covers various aspects such as brands, operating systems, battery capacity, 5G availability, internal storage, and RAM capacity. This data could be useful for understanding the mobile market trends in 2023.